load functions + ManagerTheme
