+++ manager/actions/mutate_module.dynamic.php
@@ -29,22 +29,6 @@ $tbl_site_snippets = $modx->getFullTableName('site_snippets');
 $tbl_site_templates = $modx->getFullTableName('site_templates');
 $tbl_site_tmplvars = $modx->getFullTableName('site_tmplvars');
 
-if(!function_exists('createGUID')) {
-    /**
-     * create globally unique identifiers (guid)
-     *
-     * @return string
-     */
-    function createGUID()
-    {
-        srand((double)microtime() * 1000000);
-        $r = rand();
-        $u = uniqid(getmypid() . $r . (double)microtime() * 1000000, 1);
-        $m = md5($u);
-
-        return $m;
-    }
-}
 
 // check to see the module editor isn't locked
 if($lockedEl = $modx->elementIsLocked(6, $id)) {
