+++ manager/actions/mutate_plugin.dynamic.php
@@ -58,16 +58,7 @@ $lockElementId = $id;
 $lockElementType = 5;
 require_once(MODX_MANAGER_PATH . 'includes/active_user_locks.inc.php');
 
-if(!function_exists('bold')) {
-    /**
-     * @param bool $cond
-     * @return string
-     */
-    function bold($cond = false)
-    {
-        return ($cond !== false) ? ' style="background-color:#777;color:#fff;"' : '';
-    }
-}
+
 
 ?>
 <script language="JavaScript">
@@ -684,14 +675,6 @@ if(!function_exists('bold')) {
                     echoEventRows($evtnames);
                 }
 
-                if(!function_exists('echoEventRows')) {
-                    function echoEventRows(&$evtnames)
-                    {
-                        echo '<div class="row form-row"><div class="col-sm-6 col-md-4 col-lg-3">' . implode('</div><div class="col-sm-6 col-md-4 col-lg-3">',
-                                $evtnames) . '</div></div>';
-                        $evtnames = array();
-                    }
-                }
 
                 ?>
             </div>
