+++ manager/actions/resources/mgrResources.class.php
@@ -3,4 +3,7 @@ if( ! defined('IN_MANAGER_MODE') || IN_MANAGER_MODE !== true) {
     die("<b>INCLUDE_ORDERING_ERROR</b><br /><br />Please use the EVO Content Manager instead of accessing this file directly.");
 }
 
+/**
+ * Don't delete
+ */
 class mgrResources extends EvolutionCMS\Legacy\mgrResources{}
