+++ manager/actions/welcome.static.php
@@ -31,52 +31,53 @@ if($modx->hasPermission('messages')) {
 	$ph['MessageInfo'] = implode("\n", $msg);
 }
 
+$iconTpl = $managerTheme->getPartial('welcome/WrapIcon');
 // setup icons
 if($modx->hasPermission('new_user') || $modx->hasPermission('edit_user')) {
 	$icon = '<i class="[&icons_security_large&]" alt="[%user_management_title%]"> </i>[%user_management_title%]';
-	$ph['SecurityIcon'] = wrapIcon($icon, 75);
+	$ph['SecurityIcon'] = sprintf($iconTpl,$icon, 75);
 }
 if($modx->hasPermission('new_web_user') || $modx->hasPermission('edit_web_user')) {
 	$icon = '<i class="[&icons_webusers_large&]" alt="[%web_user_management_title%]"> </i>[%web_user_management_title%]';
-	$ph['WebUserIcon'] = wrapIcon($icon, 99);
+	$ph['WebUserIcon'] = sprintf($iconTpl,$icon, 99);
 }
 if($modx->hasPermission('new_module') || $modx->hasPermission('edit_module')) {
 	$icon = '<i class="[&icons_modules_large&]" alt="[%manage_modules%]"> </i>[%modules%]';
-	$ph['ModulesIcon'] = wrapIcon($icon, 106);
+	$ph['ModulesIcon'] = sprintf($iconTpl,$icon, 106);
 }
 if($modx->hasPermission('new_template') || $modx->hasPermission('edit_template') || $modx->hasPermission('new_snippet') || $modx->hasPermission('edit_snippet') || $modx->hasPermission('new_plugin') || $modx->hasPermission('edit_plugin') || $modx->hasPermission('manage_metatags')) {
 	$icon = '<i class="[&icons_resources_large&]" alt="[%element_management%]"> </i>[%elements%]';
-	$ph['ResourcesIcon'] = wrapIcon($icon, 76);
+	$ph['ResourcesIcon'] = sprintf($iconTpl,$icon, 76);
 }
 if($modx->hasPermission('bk_manager')) {
 	$icon = '<i class="[&icons_backup_large&]" alt="[%bk_manager%]"> </i>[%backup%]';
-	$ph['BackupIcon'] = wrapIcon($icon, 93);
+	$ph['BackupIcon'] = sprintf($iconTpl,$icon, 93);
 }
 if($modx->hasPermission('help')) {
 	$icon = '<i class="[&icons_help_large&]" alt="[%help%]" /> </i>[%help%]';
-	$ph['HelpIcon'] = wrapIcon($icon, 9);
+	$ph['HelpIcon'] = sprintf($iconTpl,$icon, 9);
 }
 
 if($modx->hasPermission('new_document')) {
 	$icon = '<i class="[&icons_resource_large&]"></i>[%add_resource%]';
-	$ph['ResourceIcon'] = wrapIcon($icon, 4);
+	$ph['ResourceIcon'] = sprintf($iconTpl,$icon, 4);
 	$icon = '<i class="[&icons_weblink_large&]"></i>[%add_weblink%]';
-	$ph['WeblinkIcon'] = wrapIcon($icon, 72);
+	$ph['WeblinkIcon'] = sprintf($iconTpl,$icon, 72);
 }
 if($modx->hasPermission('assets_images')) {
 	$icon = '<i class="[&icons_images_large&]"></i>[%images_management%]';
-	$ph['ImagesIcon'] = wrapIcon($icon, 72);
+	$ph['ImagesIcon'] = sprintf($iconTpl,$icon, 72);
 }
 if($modx->hasPermission('assets_files')) {
 	$icon = '<i class="[&icons_files_large&]"></i>[%files_management%]';
-	$ph['FilesIcon'] = wrapIcon($icon, 72);
+	$ph['FilesIcon'] = sprintf($iconTpl,$icon, 72);
 }
 if($modx->hasPermission('change_password')) {
 	$icon = '<i class="[&icons_password_large&]"></i>[%change_password%]';
-	$ph['PasswordIcon'] = wrapIcon($icon, 28);
+	$ph['PasswordIcon'] = sprintf($iconTpl,$icon, 28);
 }
 $icon = '<i class="[&icons_logout_large&]"></i>[%logout%]';
-$ph['LogoutIcon'] = wrapIcon($icon, 8);
+$ph['LogoutIcon'] = sprintf($iconTpl,$icon, 8);
 
 // do some config checks
 if(($modx->config['warning_visibility'] == 0 && $_SESSION['mgrRole'] == 1) || $modx->config['warning_visibility'] == 1) {
@@ -121,7 +122,8 @@ if($count > 1) {
 }*/
 $ph['show_multiple_sessions'] = 'none';
 
-$ph['RecentInfo'] = getRecentInfo();
+$modx->addSnippet('recentInfoList', $managerTheme->getPartial('welcome/RecentInfo/Snippet'));
+$ph['RecentInfo'] = $managerTheme->getPartial('welcome/RecentInfo/Template');
 
 $tpl = '
 <table class="table data">
@@ -248,7 +250,7 @@ $ph['modx_news_title'] = $_lang['modx_news_title'];
 
 $modx->toPlaceholders($ph);
 
-$script = getStartUpScript();
+$script = $managerTheme->getPartial('welcome/StartUpScript');
 $modx->regClientScript($script);
 
 // invoke event OnManagerWelcomePrerender
@@ -395,7 +397,7 @@ usort($widgets, function ($a, $b) {
 	return $a['menuindex'] - $b['menuindex'];
 });
 
-$tpl = getTplWidget();
+$tpl = $managerTheme->getPartial('welcome/Widget');
 $output = '';
 foreach($widgets as $widget) {
 	if ($widget['hide'] != '1'){
@@ -460,209 +462,3 @@ echo $content;
 //	<a href="javascript:;" class="setting"><i class="fa fa-cog"></i></a>
 //  <a href="javascript:;" class="closed"><i class="fa fa-close"></i></a>
 //</span>
-if(!function_exists('getTplWidget')) {
-    function getTplWidget()
-    { // recent document info
-        return '
-		<div class="[+cols+]" id="[+id+]">
-			<div class="card"[+cardAttr+]>
-				<div class="card-header"[+headAttr+]> <i class="fa [+icon+]"></i> [+title+] </div>
-				<div class="card-block"[+bodyAttr+]> [+body+] </div>
-			</div>
-		</div>
-';
-    }
-}
-
-if(!function_exists('getRecentInfo')) {
-    function getRecentInfo()
-    { // recent document info
-        $modx = evolutionCMS();
-
-        $modx->addSnippet('recentInfoList', 'getRecentInfoList');
-
-        $html = '
-			<div class="table-responsive">
-				<table class="table data">
-					<thead>
-						<tr>
-							<th style="width: 1%">[%id%]</th>
-							<th>[%resource_title%]</th>
-							<th style="width: 1%">[%page_data_edited%]</th>
-							<th style="width: 1%">[%user%]</th>
-							<th style="width: 1%; text-align: center">[%mgrlog_action%]</th>
-						</tr>
-					</thead>
-					<tbody>
-					[[#recentInfoList]]
-					</tbody>
-				</table>
-			</div>
-';
-
-        return $html;
-    }
-}
-
-if(!function_exists('getRecentInfoList')) {
-    function getRecentInfoList()
-    {
-        $modx = evolutionCMS();
-
-        $rs = $modx->db->select('*', '[+prefix+]site_content', '', 'editedon DESC', 10);
-
-        if ($modx->db->getRecordCount($rs) < 1) {
-            return '<tr><td>[%no_activity_message%]</td></tr>';
-        }
-
-        $tpl = getRecentInfoRowTpl();
-
-        $btntpl['edit'] = '<a title="[%edit_resource%]" href="index.php?a=27&amp;id=[+id+]" target="main"><i class="fa fa-edit fa-fw"></i></a> ';
-        $btntpl['preview_btn'] = '<a [+preview_disabled+]" title="[%preview_resource%]" target="_blank" href="../index.php?&amp;id=[+id+]"><i class="fa fa-eye fa-fw"></i></a> ';
-
-        $output = array();
-        while ($ph = $modx->db->getRow($rs)) {
-            $docid = $ph['id'];
-            $_ = $modx->getUserInfo($ph['editedby']);
-            $ph['username'] = $_['username'];
-
-            if ($ph['deleted'] == 1) {
-                $ph['status'] = 'deleted text-danger';
-            } elseif ($ph['published'] == 0) {
-                $ph['status'] = 'unpublished font-italic text-muted';
-            } else {
-                $ph['status'] = 'published';
-            }
-
-            if ($modx->hasPermission('edit_document')) {
-                $ph['edit_btn'] = str_replace('[+id+]', $docid, $btntpl['edit']);
-            } else {
-                $ph['edit_btn'] = '';
-            }
-
-            $preview_disabled = ($ph['deleted'] == 1) ? 'disabled' : '';
-            $ph['preview_btn'] = str_replace(array(
-                '[+id+]',
-                '[+preview_disabled+]'
-            ), array(
-                $docid,
-                $preview_disabled
-            ), $btntpl['preview_btn']);
-
-            if ($modx->hasPermission('delete_document')) {
-                if ($ph['deleted'] == 0) {
-                    $delete_btn = '<a onclick="return confirm(\'[%confirm_delete_record%]\')" title="[%delete_resource%]" href="index.php?a=6&amp;id=[+id+]" target="main"><i class="fa fa-trash fa-fw"></i></a> ';
-                } else {
-                    $delete_btn = '<a onclick="return confirm(\'[%confirm_undelete%]\')" title="[%undelete_resource%]" href="index.php?a=63&amp;id=[+id+]" target="main"><i class="fa fa-arrow-circle-o-up fa-fw"></i></a> ';
-                }
-                $ph['delete_btn'] = str_replace('[+id+]', $docid, $delete_btn);
-            } else {
-                $ph['delete_btn'] = '';
-            }
-
-            if ($ph['deleted'] == 1 && $ph['published'] == 0) {
-                $publish_btn = '<a class="disabled" title="[%publish_resource%]" href="index.php?a=61&amp;id=[+id+]" target="main"><i class="fa fa-arrow-up fa-fw"></i></a> ';
-            } elseif ($ph['deleted'] == 1 && $ph['published'] == 1) {
-                $publish_btn = '<a class="disabled" title="[%publish_resource%]" href="index.php?a=61&amp;id=[+id+]" target="main"><i class="fa fa-arrow-down fa-fw"></i></a> ';
-            } elseif ($ph['deleted'] == 0 && $ph['published'] == 0) {
-                $publish_btn = '<a title="[%publish_resource%]" href="index.php?a=61&amp;id=[+id+]" target="main"><i class="fa fa-arrow-up fa-fw"></i></a> ';
-            } else {
-                $publish_btn = '<a title="[%unpublish_resource%]" href="index.php?a=62&amp;id=[+id+]" target="main"><i class="fa fa-arrow-down fa-fw"></i></a> ';
-            }
-            $ph['publish_btn'] = str_replace('[+id+]', $docid, $publish_btn);
-
-            $ph['info_btn'] = str_replace('[+id+]', $docid,
-                '<a title="[%resource_overview%]" data-toggle="collapse" data-target=".collapse[+id+]"><i class="fa fa-info fa-fw"></i></a>');
-
-            if ($ph['longtitle'] == '') {
-                $ph['longtitle'] = '(<i>[%not_set%]</i>)';
-            }
-            if ($ph['description'] == '') {
-                $ph['description'] = '(<i>[%not_set%]</i>)';
-            }
-            if ($ph['introtext'] == '') {
-                $ph['introtext'] = '(<i>[%not_set%]</i>)';
-            }
-            if ($ph['alias'] == '') {
-                $ph['alias'] = '(<i>[%not_set%]</i>)';
-            }
-
-            $output[] = $modx->parseText($tpl, $ph);
-        }
-
-        return implode("\n", $output);
-    }
-}
-
-if(!function_exists('getRecentInfoRowTpl')) {
-    function getRecentInfoRowTpl()
-    {
-        $tpl = '
-						<tr>
-							<td data-toggle="collapse" data-target=".collapse[+id+]" class="text-right"><span class="label label-info">[+id+]</span></td>
-							<td data-toggle="collapse" data-target=".collapse[+id+]"><a class="[+status+]" title="[%edit_resource%]" href="index.php?a=3&amp;id=[+id+]" target="main">[+pagetitle+]</a></td>
-							<td data-toggle="collapse" data-target=".collapse[+id+]" class="text-right text-nowrap">[+editedon:math("%s+[(server_offset_time)]"):dateFormat+]</td>
-							<td data-toggle="collapse" data-target=".collapse[+id+]">[+username+]</td>
-							<td style="text-align: right;" class="actions">[+edit_btn+][+preview_btn+][+delete_btn+][+publish_btn+][+info_btn+]</td>
-						</tr>
-						<tr class="resource-overview-accordian collapse collapse[+id+]">
-							<td colspan="6">
-								<div class="overview-body text-small">
-									<ul>
-										<li><b>[%long_title%]</b>: [+longtitle+]</li>
-										<li><b>[%description%]</b>: [+description+]</li>
-										<li><b>[%resource_summary%]</b>: [+introtext+]</li>
-										<li><b>[%type%]</b>: [+type:is(reference):then([%weblink%]):else([%resource%])+]</li>
-										<li><b>[%resource_alias%]</b>: [+alias+]</li>
-										<li><b>[%page_data_cacheable%]</b>: [+cacheable:is(1):then([%yes%]):else([%no%])+]</li>
-										<li><b>[%resource_opt_show_menu%]</b>: [+hidemenu:is(0):then([%yes%]):else([%no%])+]</li>
-										<li><b>[%page_data_template%]</b>: [+template:templatename+]</li>
-									</ul>
-								</div>
-							</td>
-						</tr>';
-
-        return $tpl;
-    }
-}
-
-if(!function_exists('wrapIcon')) {
-// setup icons
-    function wrapIcon($i, $action)
-    {
-        return sprintf('<a href="index.php?a=%s" target="main"><span class="wm_button" style="border:0">%s</span></a>',
-            $action, $i);
-    }
-}
-
-if(!function_exists('getStartUpScript')) {
-    function getStartUpScript()
-    {
-        $script = '
-        <script type="text/javascript">
-        function hideConfigCheckWarning(key) {
-        	var xhr = new XMLHttpRequest();
-        	xhr.open("POST", "index.php?a=118", true);
-        	xhr.setRequestHeader("Content-Type", "application/x-www-form-urlencoded;");
-        	xhr.onload = function() {
-        		if(this.readyState === 4) {
-        			var fieldset = document.getElementById(key + "_warning_wrapper").parentNode.parentNode;
-        			fieldset.className = "collapse";
-        		}
-        	};
-        	xhr.send("action=setsetting&key=_hide_configcheck_" + key + "&value=1");
-        }
-		(function($) {
-			$("[data-toggle=\'collapse\']").click(function(e) {
-				if(e.target.tagName === "A") return;
-				if($(this).data("target")) {
-					$($(this).data("target")).toggleClass("in")
-				}
-			});
-		})(jQuery);        
-        </script>
-';
-
-        return $script;
-    }
-}
