+++ manager/includes/src/Interfaces/ManagerThemeInterface.php
@@ -1,0 +1,3 @@
+<?php namespace EvolutionCMS\Interfaces;
+
+interface ManagerThemeInterface{}
