+++ manager/media/style/default/partials/welcome/RecentInfo/Template.tpl
@@ -1,0 +1,16 @@
+<div class="table-responsive">
+    <table class="table data">
+        <thead>
+        <tr>
+            <th style="width: 1%">[%id%]</th>
+            <th>[%resource_title%]</th>
+            <th style="width: 1%">[%page_data_edited%]</th>
+            <th style="width: 1%">[%user%]</th>
+            <th style="width: 1%; text-align: center">[%mgrlog_action%]</th>
+        </tr>
+        </thead>
+        <tbody>
+        [[#recentInfoList]]
+        </tbody>
+    </table>
+</div>
