+++ manager/media/style/default/partials/welcome/RecentInfoList/InfoBtn.tpl
@@ -1,0 +1,1 @@
+<a title="[%resource_overview%]" data-toggle="collapse" data-target=".collapse[+id+]"><i class="fa fa-info fa-fw"></i></a>
