+++ manager/media/style/default/partials/welcome/Widget.tpl
@@ -1,0 +1,6 @@
+<div class="[+cols+]" id="[+id+]">
+    <div class="card"[+cardAttr+]>
+        <div class="card-header"[+headAttr+]> <i class="fa [+icon+]"></i> [+title+] </div>
+        <div class="card-block"[+bodyAttr+]> [+body+] </div>
+    </div>
+</div>
