+++ manager/media/style/default/partials/welcome/WrapIcon.tpl
@@ -1,0 +1,1 @@
+<a href="index.php?a=%s" target="main"><span class="wm_button" style="border:0">%s</span></a>
