+++ manager/processors/execute_module.processor.php
@@ -66,49 +66,3 @@ if (strpos(trim($output),'<')===0 && strpos(trim($output),'<?xml')!==0) {
 }
 echo $output;
 include MODX_MANAGER_PATH."includes/sysalert.display.inc.php";
-
-if(!function_exists('evalModule')) {
-    /**
-     * evalModule
-     *
-     * @param string $moduleCode
-     * @param array $params
-     * @return string
-     */
-    function evalModule($moduleCode, $params)
-    {
-        $modx = evolutionCMS();
-        $modx->event->params = &$params; // store params inside event object
-        if (is_array($params)) {
-            extract($params, EXTR_SKIP);
-        }
-        ob_start();
-        $mod = eval($moduleCode);
-        $msg = ob_get_contents();
-        ob_end_clean();
-        if (isset($php_errormsg)) {
-            $error_info = error_get_last();
-            switch ($error_info['type']) {
-                case E_NOTICE :
-                    $error_level = 1;
-                case E_USER_NOTICE :
-                    break;
-                case E_DEPRECATED :
-                case E_USER_DEPRECATED :
-                case E_STRICT :
-                    $error_level = 2;
-                    break;
-                default:
-                    $error_level = 99;
-            }
-            if ($modx->config['error_reporting'] === '99' || 2 < $error_level) {
-                $modx->messageQuit('PHP Parse Error', '', true, $error_info['type'], $error_info['file'],
-                    $_SESSION['itemname'] . ' - Module', $error_info['message'], $error_info['line'], $msg);
-                $modx->event->alert("An error occurred while loading. Please see the event log for more information<p>{$msg}</p>");
-            }
-        }
-        unset($modx->event->params);
-
-        return $mod . $msg;
-    }
-}
