+++ manager/processors/login.processor.php
@@ -287,140 +287,3 @@ if($id > 0) {
 		header($header);
 	}
 }
-
-if(!function_exists('jsAlert')) {
-    /**
-     * show javascript alert
-     *
-     * @param string $msg
-     */
-    function jsAlert($msg)
-    {
-        $modx = evolutionCMS();
-        if ($_POST['ajax'] != 1) {
-            echo "<script>window.setTimeout(\"alert('" . addslashes($modx->db->escape($msg)) . "')\",10);history.go(-1)</script>";
-        } else {
-            echo $msg . "\n";
-        }
-    }
-}
-
-if(!function_exists('login')) {
-    /**
-     * @param string $username
-     * @param string $givenPassword
-     * @param string $dbasePassword
-     * @return bool
-     */
-    function login($username, $givenPassword, $dbasePassword)
-    {
-        $modx = evolutionCMS();
-
-        return $modx->phpass->CheckPassword($givenPassword, $dbasePassword);
-    }
-}
-
-if(!function_exists('loginV1')) {
-    /**
-     * @param int $internalKey
-     * @param string $givenPassword
-     * @param string $dbasePassword
-     * @param string $username
-     * @return bool
-     */
-    function loginV1($internalKey, $givenPassword, $dbasePassword, $username)
-    {
-        $modx = evolutionCMS();
-
-        $user_algo = $modx->manager->getV1UserHashAlgorithm($internalKey);
-
-        if (!isset($modx->config['pwd_hash_algo']) || empty($modx->config['pwd_hash_algo'])) {
-            $modx->config['pwd_hash_algo'] = 'UNCRYPT';
-        }
-
-        if ($user_algo !== $modx->config['pwd_hash_algo']) {
-            $bk_pwd_hash_algo = $modx->config['pwd_hash_algo'];
-            $modx->config['pwd_hash_algo'] = $user_algo;
-        }
-
-        if ($dbasePassword != $modx->manager->genV1Hash($givenPassword, $internalKey)) {
-            return false;
-        }
-
-        updateNewHash($username, $givenPassword);
-
-        return true;
-    }
-}
-
-if(!function_exists('loginMD5')) {
-    /**
-     * @param int $internalKey
-     * @param string $givenPassword
-     * @param string $dbasePassword
-     * @param string $username
-     * @return bool
-     */
-    function loginMD5($internalKey, $givenPassword, $dbasePassword, $username)
-    {
-        $modx = evolutionCMS();
-
-        if ($dbasePassword != md5($givenPassword)) {
-            return false;
-        }
-        updateNewHash($username, $givenPassword);
-
-        return true;
-    }
-}
-
-if(!function_exists('updateNewHash')) {
-    /**
-     * @param string $username
-     * @param string $password
-     */
-    function updateNewHash($username, $password)
-    {
-        $modx = evolutionCMS();
-
-        $field = array();
-        $field['password'] = $modx->phpass->HashPassword($password);
-        $modx->db->update($field, '[+prefix+]manager_users', "username='{$username}'");
-    }
-}
-
-if(!function_exists('incrementFailedLoginCount')) {
-    /**
-     * @param int $internalKey
-     * @param int $failedlogins
-     * @param int $failed_allowed
-     * @param int $blocked_minutes
-     */
-    function incrementFailedLoginCount($internalKey, $failedlogins, $failed_allowed, $blocked_minutes)
-    {
-        $modx = evolutionCMS();
-
-        $failedlogins += 1;
-
-        $fields = array('failedlogincount' => $failedlogins);
-        if ($failedlogins >= $failed_allowed) //block user for too many fail attempts
-        {
-            $fields['blockeduntil'] = time() + ($blocked_minutes * 60);
-        }
-
-        $modx->db->update($fields, '[+prefix+]user_attributes', "internalKey='{$internalKey}'");
-
-        if ($failedlogins < $failed_allowed) {
-            //sleep to help prevent brute force attacks
-            $sleep = (int)$failedlogins / 2;
-            if ($sleep > 5) {
-                $sleep = 5;
-            }
-            sleep($sleep);
-        }
-        @session_destroy();
-        session_unset();
-
-        return;
-    }
-}
