+++ manager/processors/save_plugin.processor.php
@@ -169,80 +169,3 @@ switch ($_POST['mode']) {
     default:
         $modx->webAlertAndQuit('No operation set in request.');
 }
-
-if(!function_exists('saveEventListeners')) {
-# Save Plugin Event Listeners
-    function saveEventListeners($id, $sysevents, $mode)
-    {
-        $modx = evolutionCMS();
-        // save selected system events
-        $formEventList = array();
-        foreach ($sysevents as $evtId) {
-            if (!preg_match('@^[1-9][0-9]*$@', $evtId)) {
-                $evtId = getEventIdByName($evtId);
-            }
-            if ($mode == '101') {
-                $rs = $modx->db->select('max(priority) as priority', '[+prefix+]site_plugin_events',
-                    "evtid='{$evtId}'");
-            } else {
-                $rs = $modx->db->select('priority', '[+prefix+]site_plugin_events',
-                    "evtid='{$evtId}' and pluginid='{$id}'");
-            }
-            $prevPriority = $modx->db->getValue($rs);
-            if ($mode == '101') {
-                $priority = isset($prevPriority) ? $prevPriority + 1 : 1;
-            } else {
-                $priority = isset($prevPriority) ? $prevPriority : 1;
-            }
-            $formEventList[] = array('pluginid' => $id, 'evtid' => $evtId, 'priority' => $priority);
-        }
-
-        $evtids = array();
-        foreach ($formEventList as $eventInfo) {
-            $where = vsprintf("pluginid='%s' AND evtid='%s'", $eventInfo);
-            $modx->db->save($eventInfo, '[+prefix+]site_plugin_events', $where);
-            $evtids[] = $eventInfo['evtid'];
-        }
-
-        $rs = $modx->db->select('*', '[+prefix+]site_plugin_events', sprintf("pluginid='%s'", $id));
-        $dbEventList = array();
-        $del = array();
-        while ($row = $modx->db->getRow($rs)) {
-            if (!in_array($row['evtid'], $evtids)) {
-                $del[] = $row['evtid'];
-            }
-        }
-
-        if (empty($del)) {
-            return;
-        }
-
-        foreach ($del as $delid) {
-            $modx->db->delete('[+prefix+]site_plugin_events', sprintf("evtid='%s' AND pluginid='%s'", $delid, $id));
-        }
-    }
-}
-
-if(!function_exists('getEventIdByName')) {
-    /**
-     * @param string $name
-     * @return string|int
-     */
-    function getEventIdByName($name)
-    {
-        $modx = evolutionCMS();
-        static $eventIds = array();
-
-        if (isset($eventIds[$name])) {
-            return $eventIds[$name];
-        }
-
-        $rs = $modx->db->select('id, name', '[+prefix+]system_eventnames');
-        while ($row = $modx->db->getRow($rs)) {
-            $eventIds[$row['name']] = $row['id'];
-        }
-
-        return $eventIds[$name];
-    }
-}
-
