+++ manager/processors/save_template.processor.php
@@ -143,42 +143,3 @@ switch ($_POST['mode']) {
     default:
         $modx->webAlertAndQuit("No operation set in request.");
 }
-
-if(!function_exists('saveTemplateAccess')) {
-    /**
-     * @param int $id
-     */
-    function saveTemplateAccess($id)
-    {
-        $modx = evolutionCMS();
-        if ($_POST['tvsDirty'] == 1) {
-            $newAssignedTvs = $_POST['assignedTv'];
-
-            // Preserve rankings of already assigned TVs
-            $rs = $modx->db->select("tmplvarid, rank", $modx->getFullTableName('site_tmplvar_templates'),
-                "templateid='{$id}'", "");
-
-            $ranksArr = array();
-            $highest = 0;
-            while ($row = $modx->db->getRow($rs)) {
-                $ranksArr[$row['tmplvarid']] = $row['rank'];
-                $highest = $highest < $row['rank'] ? $row['rank'] : $highest;
-            };
-
-            $modx->db->delete($modx->getFullTableName('site_tmplvar_templates'), "templateid='{$id}'");
-            if (empty($newAssignedTvs)) {
-                return;
-            }
-            foreach ($newAssignedTvs as $tvid) {
-                if (!$id || !$tvid) {
-                    continue;
-                }    // Dont link zeros
-                $modx->db->insert(array(
-                    'templateid' => $id,
-                    'tmplvarid'  => $tvid,
-                    'rank'       => isset($ranksArr[$tvid]) ? $ranksArr[$tvid] : $highest += 1 // append TVs to rank
-                ), $modx->getFullTableName('site_tmplvar_templates'));
-            }
-        }
-    }
-}
