+++ manager/processors/save_user.processor.php
@@ -45,25 +45,25 @@ $user_groups = $input['user_groups'];
 
 // verify password
 if($passwordgenmethod == "spec" && $input['specifiedpassword'] != $input['confirmpassword']) {
-	webAlertAndQuit("Password typed is mismatched");
+	webAlertAndQuit("Password typed is mismatched", 12);
 }
 
 // verify email
 if($email == '' || !preg_match("/^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,24}$/i", $email)) {
-	webAlertAndQuit("E-mail address doesn't seem to be valid!");
+	webAlertAndQuit("E-mail address doesn't seem to be valid!", 12);
 }
 
 // verify admin security
 if($_SESSION['mgrRole'] != 1) {
 	// Check to see if user tried to spoof a "1" (admin) role
 	if(!$modx->hasPermission('save_role')) {
-		webAlertAndQuit("Illegal attempt to create/modify administrator by non-administrator!");
+		webAlertAndQuit("Illegal attempt to create/modify administrator by non-administrator!", 12);
 	}
 	// Verify that the user being edited wasn't an admin and the user ID got spoofed
 	$rs = $modx->db->select('count(internalKey)', $tbl_user_attributes, "internalKey='{$id}' AND role=1");
 	$limit = $modx->db->getValue($rs);
 	if($limit > 0) {
-		webAlertAndQuit("You cannot alter an administrative user.");
+		webAlertAndQuit("You cannot alter an administrative user.", 12);
 	}
 
 }
@@ -74,29 +74,29 @@ switch($input['mode']) {
 		$rs = $modx->db->select('count(id)', $tbl_manager_users, sprintf("username='%s'", $modx->db->escape($newusername)));
 		$limit = $modx->db->getValue($rs);
 		if($limit > 0) {
-			webAlertAndQuit("User name is already in use!");
+			webAlertAndQuit("User name is already in use!", 12);
 		}
 
 		// check if the email address already exist
 		$rs = $modx->db->select('count(internalKey)', $tbl_user_attributes, sprintf("email='%s' AND id!='%s'", $modx->db->escape($email), $id));
 		$limit = $modx->db->getValue($rs);
 		if($limit > 0) {
-			webAlertAndQuit("Email is already in use!");
+			webAlertAndQuit("Email is already in use!", 12);
 		}
 
 		// generate a new password for this user
 		if($specifiedpassword != "" && $passwordgenmethod == "spec") {
 			if(strlen($specifiedpassword) < 6) {
-				webAlertAndQuit("Password is too short!");
+				webAlertAndQuit("Password is too short!", 12);
 			} else {
 				$newpassword = $specifiedpassword;
 			}
 		} elseif($specifiedpassword == "" && $passwordgenmethod == "spec") {
-			webAlertAndQuit("You didn't specify a password for this user!");
+			webAlertAndQuit("You didn't specify a password for this user!", 12);
 		} elseif($passwordgenmethod == 'g') {
 			$newpassword = generate_password(8);
 		} else {
-			webAlertAndQuit("No password generation method specified!");
+			webAlertAndQuit("No password generation method specified!", 12);
 		}
 
 		// invoke OnBeforeUserFormSave event
@@ -116,7 +116,7 @@ switch($input['mode']) {
 		$modx->db->insert($field, $tbl_user_attributes);
 
 		// Save user settings
-		saveUserSettings($internalKey);
+        saveManagerUserSettings($internalKey);
 
 		// invoke OnManagerSaveUser event
 		$modx->invokeEvent("OnManagerSaveUser", array(
@@ -154,7 +154,7 @@ switch($input['mode']) {
 		// end of user_groups stuff!
 
 		if($passwordnotifymethod == 'e') {
-			sendMailMessage($email, $newusername, $newpassword, $fullname);
+            sendMailMessageForUser($email, $newusername, $newpassword, $fullname, $signupemail_message, MODX_MANAGER_URL);
 			if($input['stay'] != '') {
 				$a = ($input['stay'] == '2') ? "12&id={$internalKey}" : "11";
 				$header = "Location: index.php?a={$a}&r=2&stay=" . $input['stay'];
@@ -202,34 +202,34 @@ switch($input['mode']) {
 		if($genpassword == 1) {
 			if($specifiedpassword != "" && $passwordgenmethod == "spec") {
 				if(strlen($specifiedpassword) < 6) {
-					webAlertAndQuit("Password is too short!");
+					webAlertAndQuit("Password is too short!", 12);
 				} else {
 					$newpassword = $specifiedpassword;
 				}
 			} elseif($specifiedpassword == "" && $passwordgenmethod == "spec") {
-				webAlertAndQuit("You didn't specify a password for this user!");
+				webAlertAndQuit("You didn't specify a password for this user!", 12);
 			} elseif($passwordgenmethod == 'g') {
 				$newpassword = generate_password(8);
 			} else {
-				webAlertAndQuit("No password generation method specified!");
+				webAlertAndQuit("No password generation method specified!", 12);
 			}
 		}
 		if($passwordnotifymethod == 'e') {
-			sendMailMessage($email, $newusername, $newpassword, $fullname);
+            sendMailMessageForUser($email, $newusername, $newpassword, $fullname, $signupemail_message, MODX_MANAGER_URL);
 		}
 
 		// check if the username already exist
 		$rs = $modx->db->select('count(id)', $tbl_manager_users, sprintf("username='%s' AND id!='%s'", $modx->db->escape($newusername), $id));
 		$limit = $modx->db->getValue($rs);
 		if($limit > 0) {
-			webAlertAndQuit("User name is already in use!");
+			webAlertAndQuit("User name is already in use!", 12);
 		}
 
 		// check if the email address already exists
 		$rs = $modx->db->select('count(internalKey)', $tbl_user_attributes, sprintf("email='%s' AND internalKey!='%s'", $modx->db->escape($email), $id));
 		$limit = $modx->db->getValue($rs);
 		if($limit > 0) {
-			webAlertAndQuit("Email is already in use!");
+			webAlertAndQuit("Email is already in use!", 12);
 		}
 
 		// invoke OnBeforeUserFormSave event
@@ -250,7 +250,7 @@ switch($input['mode']) {
 		$modx->db->update($field, $tbl_user_attributes, "internalKey='{$id}'");
 
 		// Save user settings
-		saveUserSettings($id);
+        saveManagerUserSettings($id);
 
 		// Set the item name for logger
 		$_SESSION['itemname'] = $newusername;
@@ -345,173 +345,5 @@ switch($input['mode']) {
 		}
 		break;
 	default:
-		webAlertAndQuit("No operation set in request.");
-}
-
-if(!function_exists('sendMailMessage')) {
-    /**
-     * Send an email to the user
-     *
-     * @param string $email
-     * @param string $uid
-     * @param string $pwd
-     * @param string $ufn
-     */
-    function sendMailMessage($email, $uid, $pwd, $ufn)
-    {
-        $modx = evolutionCMS();
-        global $_lang, $signupemail_message;
-        global $emailsubject, $emailsender;
-        global $site_name;
-        $manager_url = MODX_MANAGER_URL;
-        $message = sprintf($signupemail_message, $uid, $pwd); // use old method
-        // replace placeholders
-        $message = str_replace("[+uid+]", $uid, $message);
-        $message = str_replace("[+pwd+]", $pwd, $message);
-        $message = str_replace("[+ufn+]", $ufn, $message);
-        $message = str_replace("[+sname+]", $site_name, $message);
-        $message = str_replace("[+saddr+]", $emailsender, $message);
-        $message = str_replace("[+semail+]", $emailsender, $message);
-        $message = str_replace("[+surl+]", $manager_url, $message);
-
-        $param = array();
-        $param['from'] = "{$site_name}<{$emailsender}>";
-        $param['subject'] = $emailsubject;
-        $param['body'] = $message;
-        $param['to'] = $email;
-        $param['type'] = 'text';
-        $rs = $modx->sendmail($param);
-        if (!$rs) {
-            $modx->manager->saveFormValues();
-            $modx->messageQuit("{$email} - {$_lang['error_sending_email']}");
-        }
-    }
-}
-
-if(!function_exists('saveUserSettings')) {
-    /**
-     * Save User Settings
-     *
-     * @param int $id
-     */
-    function saveUserSettings($id)
-    {
-        $modx = evolutionCMS();
-        $tbl_user_settings = $modx->getFullTableName('user_settings');
-
-        $ignore = array(
-            'id',
-            'oldusername',
-            'oldemail',
-            'newusername',
-            'fullname',
-            'newpassword',
-            'newpasswordcheck',
-            'passwordgenmethod',
-            'passwordnotifymethod',
-            'specifiedpassword',
-            'confirmpassword',
-            'email',
-            'phone',
-            'mobilephone',
-            'fax',
-            'dob',
-            'country',
-            'street',
-            'city',
-            'state',
-            'zip',
-            'gender',
-            'photo',
-            'comment',
-            'role',
-            'failedlogincount',
-            'blocked',
-            'blockeduntil',
-            'blockedafter',
-            'user_groups',
-            'mode',
-            'blockedmode',
-            'stay',
-            'save',
-            'theme_refresher'
-        );
-
-        // determine which settings can be saved blank (based on 'default_{settingname}' POST checkbox values)
-        $defaults = array(
-            'upload_images',
-            'upload_media',
-            'upload_flash',
-            'upload_files'
-        );
-
-        // get user setting field names
-        $settings = array();
-        foreach ($_POST as $n => $v) {
-            if (in_array($n, $ignore) || (!in_array($n, $defaults) && is_scalar($v) && trim($v) == '') || (!in_array($n,
-                        $defaults) && is_array($v) && empty($v))) {
-                continue;
-            } // ignore blacklist and empties
-            $settings[$n] = $v; // this value should be saved
-        }
-
-        foreach ($defaults as $k) {
-            if (isset($settings['default_' . $k]) && $settings['default_' . $k] == '1') {
-                unset($settings[$k]);
-            }
-            unset($settings['default_' . $k]);
-        }
-
-        $modx->db->delete($tbl_user_settings, "user='{$id}'");
-
-        foreach ($settings as $n => $vl) {
-            if (is_array($vl)) {
-                $vl = implode(",", $vl);
-            }
-            if ($vl != '') {
-                $f = array();
-                $f['user'] = $id;
-                $f['setting_name'] = $n;
-                $f['setting_value'] = $vl;
-                $f = $modx->db->escape($f);
-                $modx->db->insert($f, $tbl_user_settings);
-            }
-        }
-    }
-}
-
-if(!function_exists('webAlertAndQuit')) {
-    /**
-     * Web alert -  sends an alert to web browser
-     *
-     * @param $msg
-     */
-    function webAlertAndQuit($msg)
-    {
-        global $id, $modx;
-        $mode = $_POST['mode'];
-        $modx->manager->saveFormValues($mode);
-        $modx->webAlertAndQuit($msg, "index.php?a={$mode}" . ($mode == '12' ? "&id={$id}" : ''));
-    }
-}
-
-if(!function_exists('generate_password')) {
-    /**
-     * Generate password
-     *
-     * @param int $length
-     * @return string
-     */
-    function generate_password($length = 10)
-    {
-        $allowable_characters = "abcdefghjkmnpqrstuvxyzABCDEFGHJKLMNPQRSTUVWXYZ23456789";
-        $ps_len = strlen($allowable_characters);
-        mt_srand((double)microtime() * 1000000);
-        $pass = "";
-        for ($i = 0; $i < $length; $i++) {
-            $pass .= $allowable_characters[mt_rand(0, $ps_len - 1)];
-        }
-
-        return $pass;
-    }
+		webAlertAndQuit("No operation set in request.", 12);
 }
