+++ manager/processors/undelete_content.processor.php
@@ -43,29 +43,7 @@ if(!$deltime) {
 
 $children = array();
 
-if(!function_exists('getChildren')) {
-    /**
-     * @param int $parent
-     */
-    function getChildren($parent)
-    {
-
-        $modx = evolutionCMS();
-        global $children;
-        global $deltime;
-
-        $rs = $modx->db->select('id', $modx->getFullTableName('site_content'),
-            "parent='" . (int)$parent . "' AND deleted=1 AND deletedon='" . (int)$deltime . "'");
-        // the document has children documents, we'll need to delete those too
-        while ($row = $modx->db->getRow($rs)) {
-            $children[] = $row['id'];
-            getChildren($row['id']);
-            //echo "Found childNode of parentNode $parent: ".$row['id']."<br />";
-        }
-    }
-}
-
-getChildren($id);
+getChildrenForUnDelete($id);
 
 if(count($children)>0) {
 	$modx->db->update(
